🚚 Move `canvasHandlers` to `lib` directory
